backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
google-adk>=0.3.0
google-genai>=0.8.0
pinecone>=3.0.0
pydantic>=2.0.0
typing_extensions>=4.9.0
